Refactor versionChecker.py -> version_checker.py
