+++ medusa/server/api/v1/core.py
@@ -56,7 +56,7 @@ from ....show.History import History
 from ....show.Show import Show
 from ....system.Restart import Restart
 from ....system.Shutdown import Shutdown
-from ....versionChecker import CheckVersion
+from ....version_checker import CheckVersion
 
 indexer_ids = ["indexerid", "tvdbid"]
