+++ pytest.ini
@@ -190,7 +190,7 @@ flake8-ignore =
     medusa/traktChecker.py D100 D101 D102 D200 D202 D205 D208 D400 D401 E261 E303 E501 N802 N803 N806
     medusa/tv_cache.py D100 D101 D102 D200 D205 D400 D401 E501 F401 N802 N803 N806
     medusa/ui.py D100 D101 D102 D200 D202 D204 D205 D400 D401 E501 N802 N803 N806
-    medusa/versionChecker.py D100 D101 D102 D200 D202 D205 D400 D401 E402 E501 N802 N803 N806
+    medusa/version_checker.py D100 D101 D102 D200 D202 D205 D400 D401 E402 E501 N802 N803 N806
     medusa/helper/common.py D100 D200 D202 D205 D400 E128 E302 E501 W391
     medusa/helper/encoding.py D100 D202 D400 D401
     medusa/helper/exceptions.py D100 D200 D202 D205 D400
